refactor(agent-runtime): extract core contracts into corvus-traits (#485)
