+++ clients/agent-runtime/Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = [".", "crates/robot-kit"]
+members = [".", "crates/corvus-traits", "crates/robot-kit"]
 resolver = "2"
 
 [package]
@@ -84,6 +84,7 @@ parking_lot = "0.12"
 
 # Async traits
 async-trait = "0.1"
+corvus-traits = { path = "crates/corvus-traits" }
 
 # WASM execution support for runtime sandboxing
 wasmi = { version = "1.0", optional = true, default-features = false, features = ["std"] }
