+++ clients/agent-runtime/crates/corvus-traits/src/testing.rs
@@ -1,0 +1,5 @@
+//! Scaffolding for future shared contract-compliance tests.
+
+/// Placeholder harness type for follow-up compliance testing work.
+#[derive(Debug, Default, Clone, Copy)]
+pub struct TraitHarness;
